+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Version {next}
+  - Add a basic search feature using Laravel Scout and Algolia (issue#38)
+
 ## Version 0.0.16.3 (2016-11-25)
   - StyleCI fix